I MADE IT SO THE BUTTONS CHANGE DEPENDING ON THE TIME OF DAY. I CLARIFIED WHAT VISITING DOES. I ADDED A SKIP BUTTON. I CHANGED THE TEXT INTO A TEXT FIELD.

+++ MafiaViews/PlayingViewModel.swift
@@ -18,15 +18,20 @@ struct Player : Identifiable {
 class Game {
     var time : GameLabelState = .day
     var currentTeam : TeamText = .mafia
+    var players : [Player] = [Player("Joe"),Player("Sam"), Player("Not Killer", .mafia), Player("Yolo"), Player("Funny")]
 }
 class PlayingViewModel : ObservableObject {
-    var time : GameLabelState = .day
-    var currentRole : RoleText = .mafia
+    
     var gameModel : Game = Game()
-    var players : [Player] = [Player("Joe"),Player("Sam"), Player("Not Killer", .mafia), Player("Yolo"), Player("Funny")]
     var playerInt : Int? = 1
     var player : Player? {
-        players[playerInt ?? 0]
+        gameModel.players[playerInt ?? 0]
+    }
+    var showPlayers: [Player] {
+        gameModel.players.filter { $0.id != player?.id ?? UUID() }
+    }
+    var time : GameLabelState {
+        gameModel.time
     }
     var timeText : String {
         gameModel.time.rawValue
@@ -55,31 +60,46 @@ class PlayingViewModel : ObservableObject {
         player?.role
     }
     var rolePowers : String {
-        switch role {
-        case .mafia:
-            "Whoever you pick, you will kill unless they are saved"
-        case .townsperson:
-            "You have no powers. Visiting has no affect."
-        case .detective:
-            "Whoever you pick you learn the role of"
-        case .doctor:
-            "Whoever you pick you will be protected"
-        case nil :
-            "No action"
+        if gameModel.time == .day {
+            "Pick someone to vote to eliminate or abstain."
+        }
+        else {
+            switch role {
+            case .mafia:
+                "Whoever you pick, you will kill unless they are saved"
+            case .townsperson:
+                "You have no powers. Visiting has no affect."
+            case .detective:
+                "Whoever you pick you learn the role of"
+            case .doctor:
+                "Whoever you pick you will be protected"
+            case nil :
+                "No action"
+            }
         }
     }
     var actionText : String {
-        switch role {
-        case .mafia:
-            "Kill"
-        case .townsperson:
-            "Visit"
-        case .detective:
-            "Investigate"
-        case .doctor:
-            "Protect"
-        case nil :
-            "No action text"
+        if gameModel.time == .day {
+            "Vote"
+        }else {
+            switch role {
+            case .mafia:
+                "Kill"
+            case .townsperson:
+                "Visit"
+            case .detective:
+                "Investigate"
+            case .doctor:
+                "Protect"
+            case nil :
+                "No action text"
+            }
         }
+    }
+    init(time: GameLabelState = .night,  gameModel: Game = Game(), playerInt: Int? = nil) {
+        self.gameModel = gameModel
+        self.playerInt = playerInt
+        gameModel.time = time
+       
     }
 }
